# deploy/02. Backend/ms-auth-security/src/main/resources/static/openapi/openapi-ms-auth-security.yaml
openapi: 3.0.3
info:
  title: ms-auth-security
  description: >
    # Resumen

    Microservicio **ms-auth-security** (puerto típico **9001**): autenticación JWT, recuperación de contraseña,
    **RBAC** bajo `/api/auth/seguridad`, y envoltorio estándar **`ApiResponse<T>`** (`pe.restaurant.common.dto`).

    # Alcance de este OAS

    Documenta **`/api/auth/**`**. Las rutas **`/api/admin/**`** y **`/api/internal/**`** se describen en el contrato
    markdown del repositorio (`CONTRATO_API_MS-AUTH-SECURITY.md`).

    # Recursos

    * YAML estático: `GET /openapi/openapi-ms-auth-security.yaml`
    * Springdoc: `/api/auth/v3/api-docs`, Swagger UI `/api/auth/swagger-ui.html`

  contact:
    name: Restaurant.pe — Plataforma ERP
    email: soporte@restaurant.pe
  license:
    name: Propietario — uso interno proyecto Restaurant.pe
    url: https://restaurant.pe
  version: 2.1.0
  x-domain: Restaurant.pe / Plataforma ERP
  x-capability: Autenticación y seguridad (auth + RBAC)
  x-capability-acronym: mash
servers:
  - description: ms-auth-security (directo)
    url: http://localhost:9001
  - description: API Gateway (prefijo /api/auth/**)
    url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (JWT temporal)
      security: []
      operationId: login
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-login'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/empresas:
    get:
      tags: [Auth]
      summary: Empresas asignadas al usuario
      operationId: listarEmpresas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEmpresaUsuarioDtoList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/seleccionar-empresa:
    post:
      tags: [Auth]
      summary: Seleccionar empresa y sucursal (JWT definitivo + refresh token)
      description: >
        Con **Authorization: Bearer** (JWT temporal o definitivo) se usa el flujo clásico.
        Sin **Authorization**, envíe **email**, **password** y **passwordHash** (mismo esquema que en login)
        junto con **empresaId** y **sucursalId** para obtener JWT definitivo en un solo paso.
      operationId: seleccionarEmpresa
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
          description: Bearer JWT temporal (u opcionalmente definitivo). Omitir si envía credenciales en el cuerpo.
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-seleccion-empresa'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Cerrar sesión (invalida tokens_session si aplica)
      operationId: logout
      security: []
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
            example: Bearer <jwt>
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access y refresh (rotación)
      security: []
      operationId: refreshToken
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-refresh'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRefreshTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Perfil desde claims del access JWT
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthMeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/recuperar/obtener-email-ofuscado:
    post:
      tags: [Recuperación]
      summary: Email enmascarado por username
      security: []
      operationId: obtenerEmailOfuscado
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-recuperar'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapString'
  /api/auth/recuperar/verificar-email:
    post:
      tags: [Recuperación]
      security: []
      operationId: verificarEmailRecuperar
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-recuperar'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMapString'
  /api/auth/recuperar/enviar-codigo:
    post:
      tags: [Recuperación]
      security: []
      operationId: enviarCodigoRecuperar
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-recuperar'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/auth/recuperar/validar-codigo:
    post:
      tags: [Recuperación]
      security: []
      operationId: validarCodigoRecuperar
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-recuperar'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/auth/recuperar/cambiar-password:
    post:
      tags: [Recuperación]
      security: []
      operationId: cambiarPasswordRecuperar
      requestBody:
        $ref: '#/components/requestBodies/rq-ms-auth-recuperar'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
components:
  requestBodies:
    rq-ms-auth-login:
      description: Credenciales y metadatos opcionales de cliente para login (JWT temporal).
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginRequest'
    rq-ms-auth-seleccion-empresa:
      description: >
        Selección de empresa y sucursal para emitir JWT definitivo y refresh token.
        Si no se envía Authorization, deben incluirse email y contraseña cifrada como en login.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SeleccionEmpresaRequest'
    rq-ms-auth-refresh:
      description: Renovación de access token mediante refresh token.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RefreshTokenRequest'
    rq-ms-auth-recuperar:
      description: >
        Flujo recuperación de contraseña; los campos usados dependen del paso
        (username/email, código, nueva contraseña, etc.).
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RecuperarPasswordRequest'
...
  responses:
    BadRequest:
      description: Validación o regla de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            message: Datos de entrada inválidos o regla de negocio no cumplida
            errorCode: BAD_REQUEST
            timestamp: "03/04/2026 10:15:00"
            data: null
    Unauthorized:
      description: Token ausente, expirado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            message: No autenticado o token inválido
            errorCode: UNAUTHORIZED
            timestamp: "03/04/2026 10:15:00"
            data: null
    Forbidden:
      description: Sin permiso (empresa, admin, etc.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            message: No tiene permiso para esta operación
            errorCode: FORBIDDEN
            timestamp: "03/04/2026 10:15:00"
            data: null
...
  schemas:
    ApiResponseBase:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errorCode:
          type: string
          nullable: true
        timestamp:
          type: string
          description: dd/MM/yyyy HH:mm:ss (America/Lima)
    ApiResponseError:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              nullable: true
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              nullable: true
    ApiResponseMapString:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        passwordHash:
          type: string
        ipAddress:
          type: string
        ipPrivada:
          type: string
        browser:
          type: string
        sistemaOperativo:
          type: string
        deviceName:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          nullable: true
        tokenType:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        temporal:
          type: boolean
        userId:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        nombreCompleto:
          type: string
        empresaId:
          type: integer
          format: int64
        empresaCodigo:
          type: string
        empresaNombre:
          type: string
        empresaRuc:
          type: string
        sucursalId:
          type: integer
          format: int64
        sucursalNombre:
          type: string
        adminSistema:
          type: boolean
          description: true si auth.usuario.flag_admin_sistema = '1'
    ApiResponseLoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginResponse'
    SeleccionEmpresaRequest:
      type: object
      required: [empresaId, sucursalId]
      properties:
        empresaId:
          type: integer
          format: int64
        sucursalId:
          type: integer
          format: int64
        ipAddress:
          type: string
        ipPrivada:
          type: string
        browser:
          type: string
        sistemaOperativo:
          type: string
        email:
          type: string
          format: email
          description: Requerido si no hay Bearer (mismo uso que en login).
        password:
          type: string
          description: Contraseña cifrada (misma que en login).
        passwordHash:
          type: string
          description: Hash de verificación HMAC (misma que en login).
    EmpresaUsuarioDto:
      type: object
      properties:
        empresaId:
          type: integer
          format: int64
        codigo:
          type: string
        razonSocial:
          type: string
        ruc:
          type: string
        dbName:
          type: string
    ApiResponseEmpresaUsuarioDtoList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/EmpresaUsuarioDto'
...
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string }
        expiresInSeconds: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        email: { type: string }
        username: { type: string }
        nombres: { type: string }
        apellidos: { type: string }
        nombreCompleto: { type: string }
        empresaId: { type: integer, format: int64 }
        empresaCodigo: { type: string }
        empresaNombre: { type: string }
        empresaRuc: { type: string }
        sucursalId: { type: integer, format: int64 }
        sucursalNombre: { type: string, nullable: true }
    ApiResponseRefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RefreshTokenResponse'
    AuthMeResponse:
      type: object
      properties:
        temporal: { type: boolean }
        userId: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        nombres: { type: string }
        apellidos: { type: string }
        nombreCompleto: { type: string }
        tokensSessionId: { type: integer, format: int64, nullable: true }
        empresaId: { type: integer, format: int64, nullable: true }
        empresaCodigo: { type: string, nullable: true }
        empresaNombre: { type: string, nullable: true }
        empresaRuc: { type: string, nullable: true }
        dbName: { type: string, nullable: true }
        sucursalId: { type: integer, format: int64, nullable: true }
        sucursalNombre: { type: string, nullable: true }
        usuario:
          $ref: '#/components/schemas/MeUsuarioClaims'
        empresa:
          $ref: '#/components/schemas/MeEmpresaClaims'
        sucursal:
          $ref: '#/components/schemas/MeSucursalClaims'
    MeUsuarioClaims:
      type: object
      nullable: true
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        nombres: { type: string }
        apellidos: { type: string }
        email: { type: string }
    MeEmpresaClaims:
      type: object
      nullable: true
      properties:
        id: { type: integer, format: int64 }
        nombre: { type: string }
        ruc: { type: string }
        codigo: { type: string }
    MeSucursalClaims:
      type: object
      nullable: true
      properties:
        id: { type: integer, format: int64 }
        nombre: { type: string, nullable: true }
        ciudad: { type: string, nullable: true }
    ApiResponseAuthMeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthMeResponse'
    RecuperarPasswordRequest:
      type: object
      properties:
        username: { type: string }
        email: { type: string, format: email }
        codigo: { type: string }
        nuevaPassword: { type: string }
        nuevaPasswordHash: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Cabecera `Authorization: Bearer <token>`. El token es JWT temporal (post-login)
        o definitivo (post-selección de empresa / refresh), según la operación.